implemented decimal support

+++ API/core/models/guilds.py
@@ -17,8 +17,8 @@ class Guild(models.Model):
     manager_role_id = models.CharField(max_length=255, blank=True, null=True)
     withdrawal_address = models.CharField(max_length=64, blank=True, null=True)
 
-    total_fee_collected = models.IntegerField(default=0)
-    guild_balance = models.IntegerField(default=0)
+    total_fee_collected = models.BigIntegerField(default=0)
+    guild_balance = models.BigIntegerField(default=0)
 
     def __str__(self):
         return f"Server: {self.guild_id}; Fees: {self.total_fee_collected}"
